fix default dict import after extraction
