+++ tests/pytorch_tests/tpc_pytorch.py
@@ -14,7 +14,7 @@
 # ==============================================================================
 
 import model_compression_toolkit as mct
-from model_compression_toolkit.core.common.defaultdict import DefaultDict
+from model_compression_toolkit.defaultdict import DefaultDict
 from model_compression_toolkit.core.pytorch.constants import BIAS
 from model_compression_toolkit.target_platform_capabilities.constants import KERNEL_ATTR, BIAS_ATTR, PYTORCH_KERNEL
